Add max-depth option to limit the depth of a graph
The recursion into dependencies will stop after the requested depth has
been reached. Packages for which dependencies are not included are
automatically marked as 'end' packages.

This also improves the display as 'end' packages of Suggested dependencies
and 'end' packages that part of alternatives. Conflicts, if they have no
other relationships,  are not marked as 'end' packages.