Handle packages on skip list in alternatives
Use case is packages depending on libc6-dev|libc-dev, e.g. libslang2-dev.

Handling is quite complex as we need to allow for the fact that the
alternatives can contain a mixture of packages that should be skipped
and that should not be skipped and we have also to take into account
the installation status.