Correctly show other providers for provided virtual package
The previous commit resulted in a change in the usage3 examples: other
providers for ident-server were no longer shown. That they were shown
in the first place was a bug (the result of there being a conflict and
the virtual package previously not being marked done), but it is nice if
there are other providers.

This commit ensures they are displayed correctly, and only if there really
are other providers. It also fixes a bug where provides arrows could be
shown twice.