Use PkgApt for reverse dependencies
In test2 there are differences due to the fact that we now do show
unofficial packages installed on the system, which is a good thing.
New in the graph are debtree and flashplugin-nonfree; the numer of
rdeps for debconf is increased by 1 for both PD and D.

There's a small change in the usage4 test. Probably because libept0 gets
processed before aptitude, tasksel is now first processed as a "level 3"
instead of a level2 and thus we do not recurse to tasksel-data. And
the same increase for debconf as in test2.

For test4 there are only ordering changes.

There are remaining issues, but those are not regressions. They show most
clearly when the output of 'debtree -R --max-depth=0 debconf' is compared
for the old and new versions.