Selectively use AptPkg for build dependencies
AptPkg does not expose architecture conditions, but it does take them
into account when determining which dependencies to return. So we can
use AptPkg to show build dependencies for a specific architecture, but
not when ARCH=all. For that case we call out to apt-cache instead.
See also #546863.

Note that $pinfo is used to contain both the output from apt_cache and as
a pointer to AptPkg data.  Not the most beautiful solution, but it works.

The function check_arch() is no longer needed as AptPkg has already done
the filtering for us.