Make debtree work for all locales
Translate AptPkg's internal dependency types to the ones used in debtree.
This is not needed for C and English, but it is for other languages.
It's only needed for the DepType array and not for BuildDepends.

The trick is that $$dep{DepType} returns a localized dependency type as
string, which (0 + $$dep{DepType}) returns the dependency type as a number.

Solution is based on how apt-rdepends handles this. Solution is probably
not the most elegant, but works.